requirements.txt
Django==3.2.7
djangorestframework==3.12.4
djangorestframework-simplejwt==5.2.2
djangorestframework-jwt==1.11.0
django-environ>=0.8,<1.0
python-dotenv
psycopg2-binary>=2.9
gunicorn
django-redis==5.0.0
